ethcore: move difficulty_to_boundary to ethash crate
