+++ miner/src/work_notify.rs
@@ -28,7 +28,7 @@ use self::ethash::SeedHashCompute;
 use self::url::Url;
 use self::hyper::header::ContentType;
 
-use ethereum_types::{H256, U256, U512};
+use ethereum_types::{H256, U256};
 use parking_lot::Mutex;
 use futures::Future;
 
@@ -67,23 +67,10 @@ impl WorkPoster {
 	}
 }
 
-/// Convert an Ethash difficulty to the target boundary. Basically just `f(x) = 2^256 / x`.
-fn difficulty_to_boundary(difficulty: &U256) -> H256 {
-	assert!(!difficulty.is_zero());
-
-	if *difficulty == U256::one() {
-		U256::max_value().into()
-	} else {
-		let d = U512::from(difficulty);
-		// d > 1, so result should never overflow 256 bits
-		U256::from((U512::one() << 256) / d).into()
-	}
-}
-
 impl NotifyWork for WorkPoster {
 	fn notify(&self, pow_hash: H256, difficulty: U256, number: u64) {
 		// TODO: move this to engine
-		let target = difficulty_to_boundary(&difficulty);
+		let target = ethash::difficulty_to_boundary(&difficulty);
 		let seed_hash = &self.seed_compute.lock().hash_block_number(number);
 		let seed_hash = H256::from_slice(&seed_hash[..]);
 		let body = format!(
